+++ test/rete/dsl/hashing_test.exs
@@ -22,7 +22,7 @@ defmodule Rete.DSL.HashingTest do
   defp alpha_code(bind_pairs) do
     bind = Map.new(bind_pairs)
 
-    Codegen.alpha_expr(:order, quoted("{:order, a, b}"), quoted("{_, a, b}"), nil, bind).code
+    Codegen.alpha_expr(__ENV__, :order, quoted("{:order, a, b}"), quoted("{_, a, b}"), nil, bind).code
   end
 
   describe "codes are deterministic" do
@@ -41,7 +41,9 @@ defmodule Rete.DSL.HashingTest do
     test "the generated bindings pattern is sorted" do
       bind = Map.new(zeta: Macro.var(:zeta, nil), alpha: Macro.var(:alpha, nil))
 
-      expr = Codegen.alpha_expr(:order, quoted("{:order, a}"), quoted("{_, a}"), nil, bind)
+      expr =
+        Codegen.alpha_expr(__ENV__, :order, quoted("{:order, a}"), quoted("{_, a}"), nil, bind)
+
       {:%{}, _, pairs} = expr.__ast__.body
 
       assert [:alpha, :zeta] == Enum.map(pairs, &elem(&1, 0))